+++ src/term_img/exceptions.py
@@ -1,3 +1,0 @@
-"""Backwards-compatibility dummy for `term_image.exceptions`"""
-
-from term_image.exceptions import *