+++ src/term_img/image.py
@@ -1,3 +1,0 @@
-"""Backwards-compatibility dummy for `term_image.image`"""
-
-from term_image.image import *