+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Removed
+- [lib] `term_image.image.TermImage`.
+- [lib] `TermImageException` and `InvalidSize` from `term_image.exceptions`.
+- [lib] Top-level package `term_img`.
 
 
 ## [0.4.0] - 2022-06-27